@@ -1,9 +1,11 @@
 module checkers-backend
 
-go 1.21 
+go 1.23.0
+
+toolchain go1.24.1
 
 require (
 	// github.com/goccy/go-json v0.10.3
-	golang.org/x/net v0.27.0
-	google.golang.org/protobuf v1.34.2
+	golang.org/x/net v0.37.0
+	google.golang.org/protobuf v1.36.5
 )
