@@ -19,7 +19,7 @@ type Player struct {
 }
 
 // pingCodec is used to PING the client
-var pingCodec = websocket.Codec{Marshal: func(v interface{}) (data []byte, payloadType byte, err error) {
+var pingCodec = websocket.Codec{Marshal: func(v any) (data []byte, payloadType byte, err error) {
 	return nil, websocket.PingFrame, nil
 }}
 
